\name{GVEMGRM-package}
\alias{GVEMGRM-package}
\alias{GVEMGRM}
\docType{package}
\title{
\packageTitle{GVEMGRM}
}
\description{
\packageDescription{GVEMGRM}
}
\details{xx
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{GVEMGRM}

Maintainer: \packageMaintainer{GVEMGRM}
}
\references{Zheng, Q. Z., Shang, L., Xu, P. F.*, Shan, N. & Gao, Z. (2025). Variational Estimation for Multidimensional Graded Response Model.
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = GVEMGRM)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{GVEMGRM}}
%\section{Documentation Index}{\packageIndices{GVEMGRM}}

\seealso{xx
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

