% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwgvemgrm.R
\name{iwgvemgrm}
\alias{iwgvemgrm}
\title{IW-GVEM for MGRM}
\usage{
iwgvemgrm(
  y,
  init_A,
  init_B,
  init_Sig,
  mu_n,
  sigma_n,
  is_sigmaknown = 0,
  seed = NULL,
  S = 10,
  M = 10,
  beta1 = 0.9,
  beta2 = 0.999,
  eta_A = 0.05,
  eta_gam = 0.005,
  eta_sig = 0.005,
  eps = 0.001,
  max_iter = 100,
  tol.para = 0.001
)
}
\arguments{
\item{y}{N*J mat, graded item responses.}

\item{init_A}{J*K mat, initial value of A, a good choice is estimated A from GVEM.}

\item{init_B}{J*R vec, initial value of B, a good choice is estimated B from GVEM.}

\item{init_Sig}{K*K mat, initial value of \eqn{\Sigma}, a good choice is estimated \eqn{\Sigma} from GVEM.}

\item{mu_n}{N*K mat, parameter \eqn{\mu_i} of Gaussian variational distribution \eqn{q_i(\theta_i)}, obtain from GVEM.}

\item{sigma_n}{K*K*N array, parameter \eqn{\Sigma_i} of Gaussian variational distribution \eqn{q_i(\theta_i)}, obtain from GVEM.}

\item{is_sigmaknown}{default is 0, meaning \eqn{\Sigma} is unknown and will be estimated; 1 means \eqn{\Sigma} is known and fixed.}

\item{seed}{random number seed, default = 627.}

\item{S}{number of Monte Carlo samples.}

\item{M}{importance weight size.}

\item{beta1}{exponential decay rates of the exponential moving averages of gradient, default is 0.9.}

\item{beta2}{exponential decay rates of the exponential moving averages of squared gradient, default is 0.999.}

\item{eta_A}{learning rate of A.}

\item{eta_gam}{learning rate of \eqn{\gamma}.}

\item{eta_sig}{learning rate of \eqn{\Sigma}.}

\item{eps}{a small positive value to ensure numerical stability, default = .001.}

\item{max_iter}{maximum number of iterations. default is 100.}

\item{tol.para}{tolerance for convergence of parameters, default is 1e-3.}
}
\value{
A list contains:
  \tabular{ll}{
    \code{new_A} \tab estimated A (slope parameters).\cr
    \code{new_B} \tab estimated B (threshold parameters).\cr
    \code{new_Sig} \tab estimated \eqn{\Sigma} (correlation of latent traits).\cr
    \code{n2lb} \tab negative twice the evidence lower bound (ELBO).\cr
    \code{t} \tab number of iterations.\cr
    \code{cpu_time} \tab computing time.\cr
  }
}
\description{
Importance-weighted Gaussian variational expectation maximization algorithm 
for estimating item parameters in multidimensional graded response model.
}
\examples{
# ------------------------------------------
# -- An example for Confirmatory Analysis --
# ------------------------------------------

library(GVEMGRM)
attach(toy_data)

# -- initial parameters --
# -- note that, the initial value of B, Sig and variational parameters are unnecessary
N <- nrow(y)
J <- ncol(y)
K <- 2

model  <- (true_A!=0)*1
init_A <- matrix(.01, nrow=J, ncol=K); init_A[model==0] <- 0

# -- call gvemgrm first and then iwgvemgrm --
gvem_output <- gvemgrm(y = y, init_A = init_A, tar_mod = model)

iwgvem_output <- iwgvemgrm(
  y = y, init_A = gvem_output$new_A, init_B = gvem_output$new_B, init_Sig = gvem_output$new_sig,
  mu_n = gvem_output$mu_n, sigma_n = gvem_output$sigma_n)
names(iwgvem_output)

}
